+++ src/main/java/com/backend/tienda/repositorys/UserRepository.java
@@ -10,11 +10,12 @@ import com.backend.tienda.models.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<Usuario_general, Long> {
+public interface UserRepository extends JpaRepository<Usuario_general, Integer> {
 	
 	Optional<Usuario_general> findByNombre(String usuario_nombre);
 
 	Optional<Usuario_general> findByCorreo(String correo);
+	
 		
 	Boolean existsByNombre(String usuario_nombre);
